Create SVG render module, rename entry points, support bz2/gzip files
