+++ tests/test_sample_svgs.py
@@ -4,7 +4,7 @@ from __future__ import print_function, absolute_import, unicode_literals
 import unittest
 from distutils import dirname
 
-from svg2rlg import svg2rlg
+from svg2rlg import file_to_rlg as convert_file
 
 BASE = dirname(__file__) + "/samples/misc/"
 
@@ -18,35 +18,33 @@ class TestSampleSvgs(unittest.TestCase):
     or which one is taking a long time.
     """
 
-    def test_airbus(self):
-        svg2rlg(BASE + "airbus.svg")
-
     def test_arcs_2(self):
-        svg2rlg(BASE + "arcs02-abs.svg")
+        convert_file(BASE + "arcs02-abs.svg")
 
     def test_arcs_2_relative(self):
-        svg2rlg(BASE + "arcs02-rel.svg")
+        convert_file(BASE + "arcs02-rel.svg")
 
     def test_car(self):
-        svg2rlg(BASE + "car.svg")
+        convert_file(BASE + "car.svg")
+
+    def test_compressed_gzip(self):
+        convert_file(BASE + "car.svg.gz")
+
+    def test_compressed_bzip2(self):
+        fn = BASE + "car.svg.bz2"
+        convert_file(fn)
 
     def test_circle_arc(self):
-        svg2rlg(BASE + "circle_arc.svg")
+        convert_file(BASE + "circle_arc.svg")
 
     def test_logo_a3(self):
-        svg2rlg(BASE + "logo_a3.svg")
+        convert_file(BASE + "logo_a3.svg")
 
     def test_newlion(self):
-        svg2rlg(BASE + "newlion.svg")
-
-    def est_py221(self):
-        svg2rlg(BASE + "python221imap.svg")
+        convert_file(BASE + "newlion.svg")
 
     def test_rl_logo(self):
-        svg2rlg(BASE + "rllogo.svg")
+        convert_file(BASE + "rllogo.svg")
 
     def test_tiger(self):
-        svg2rlg(BASE + "tiger.svg")
-
-    def test_timezones(self):
-        svg2rlg(BASE + "timezones.svg")
+        convert_file(BASE + "tiger.svg")
